Rename the Media DTO to MediaDto
